+++ test/cpp/jit/test.cpp
@@ -8,13 +8,23 @@
 // 3. Write your tests as pure functions starting with "test", like "testFoo"
 // 4. Include test_foo.h here and add it to the appropriate macro listing
 #include <test/cpp/jit/test_alias_analysis.h>
+#include <test/cpp/jit/test_argument_spec.h>
+#include <test/cpp/jit/test_autodiff.h>
 #include <test/cpp/jit/test_class_parser.h>
 #include <test/cpp/jit/test_constant_pooling.h>
+#include <test/cpp/jit/test_create_autodiff_subgraphs.h>
+#include <test/cpp/jit/test_fuser.h>
+#include <test/cpp/jit/test_graph_executor.h>
+#include <test/cpp/jit/test_ir.h>
 #include <test/cpp/jit/test_irparser.h>
+#include <test/cpp/jit/test_ivalue.h>
 #include <test/cpp/jit/test_misc.h>
 #include <test/cpp/jit/test_netdef_converter.h>
+#include <test/cpp/jit/test_subgraph_utils.h>
 
 using namespace torch::jit::script;
+using namespace torch::jit::test;
+
 namespace torch {
 namespace jit {
 #define TH_FORALL_TESTS(_)         \
@@ -57,14 +67,14 @@ namespace jit {
 
 #if defined(USE_GTEST)
 
-#define JIT_GTEST(name)  \
+#define JIT_GTEST(name) \
   TEST(JitTest, name) { \
     test##name();       \
   }
 TH_FORALL_TESTS(JIT_GTEST)
 #undef JIT_TEST
 
-#define JIT_GTEST_CUDA(name)    \
+#define JIT_GTEST_CUDA(name)   \
   TEST(JitTest, name##_CUDA) { \
     test##name();              \
   }
